refactor: rename connector auth boundary
